+++ lib/core/src/screen/window.rs
@@ -312,6 +312,7 @@ impl Window {
     #[allow(clippy::cast_sign_loss)]
     #[allow(clippy::cast_precision_loss)]
     #[allow(clippy::too_many_lines)]
+    #[allow(clippy::cognitive_complexity)]
     pub fn render(
         &self,
         buf: &Buffer,